add cast of size_t

+++ src/runtime/workbench.cpp
@@ -216,9 +216,9 @@ namespace ts {
 
             // filter value
             if (value->dtype() == FLOAT32) {
-                filter_values(value->data<float>(), value->count());
+                filter_values(value->data<float>(), size_t(value->count()));
             } else if (value->dtype() == FLOAT64) {
-                filter_values(value->data<double>(), value->count());
+                filter_values(value->data<double>(), size_t(value->count()));
             }
         }
